NeighborlyAPI/requirements.txt
azure-functions==1.2.1
azure-functions-core-tools==3.0.3160
pymongo==3.10.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.23.0
urllib3==1.25.9
visitor==0.1.3
Werkzeug==0.16.1
feedgen
